LaTeX Style file for University of Guelph Thesis
Initial import of the style file.  It extends the memoir class to add
the functionality required to create a LaTeX document that conforms to
the University of Guelphs Thesis style requirements.  At this point the
file is relatively complete however it needs testing, and there are some
components missing.  Particularly the dedication environment.