Correct the layout of the titlepage
The titlepage for the thesis id now defined using relative vertical
spacing.  It will now format correctly for both the normal and ebook
versions of the thesis.  The spacing is set such that the generated
title page will align excactly with grad studies template for a thesis
set with a 11pt or 12pt font and a title requiring 2 lones of text.  A
10pt font or a shorter title will cause the vertical spacing to be out
by a small amount.  I expect that a very long title will cause similar
problems.